refactor(react-gui): give the generic property services a folder
`genericProps.service.ts` was 518 lines with two files in `helpers/` that
existed only for it. Reading is one call and writing is three, but they sat
in one file with the eight private helpers the writes need, so finding the
one rule you wanted meant reading past all of them.

`props/` splits by what the calls do: `read.ts` dumps a node's properties,
`write.ts` writes them back, `types.ts` states the shapes, and
`selContext.ts` answers the one question the writes keep having to ask --
which molecule a selection property belongs to, since the node being edited
is not always one. `helpers/resolvePropTarget.ts` moves in as `target.ts`;
nothing outside used it.

`helpers/parseGenericProps.ts` stays where it is: the density-map panel and
the animation element inspector read it too, so it is cross-domain, which is
what `helpers/` is for.

No behaviour change; the existing generic-property tests pass with only
their import paths moved.
