+++ tritium/react-gui/src/renderer/worker/shared/calls/props.ts
@@ -14,7 +14,7 @@ import type {
   SetGenericPropArgs,
   SetGenericPropResult,
   SetGenericPropsArgs,
-} from '@renderer/worker/server/services/genericProps.service'
+} from '@renderer/worker/server/services/props/props.service'
 
 export interface PropsCalls {
   getGenericProps:            { args: GetGenericPropsArgs; result: GetGenericPropsResult }
